+++ commands/liquidity_provision_submission_test.go
@@ -1,6 +1,7 @@
 package commands_test
 
 import (
+	"fmt"
 	"testing"
 
 	"code.vegaprotocol.io/protos/commands"
@@ -27,6 +28,7 @@ func TestLiquidityProvisionSubmission(t *testing.T) {
 				CommitmentAmount: "0",
 				MarketId:         "okmarketid",
 			},
+			errString: "liquidity_provision_submission.commitment_amount (is not a valid number)",
 		},
 		{
 			lp: commandspb.LiquidityProvisionSubmission{
@@ -180,5 +182,224 @@ func TestLiquidityProvisionSubmission(t *testing.T) {
 
 		assert.Error(t, err)
 		assert.EqualError(t, err, c.errString)
+	}
+}
+
+func TestCheckLiquidityProvisionCancellation(t *testing.T) {
+	type args struct {
+		cmd *commandspb.LiquidityProvisionCancellation
+	}
+	tests := []struct {
+		name      string
+		args      args
+		wantErr   assert.ErrorAssertionFunc
+		errString string
+	}{
+		{
+			name: "Should return an error if request is nil",
+			args: args{
+				cmd: nil,
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_cancellation (is required)",
+		},
+		{
+			name: "Should return an error if id is not provided",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionCancellation{
+					Id:       "",
+					MarketId: "abcd",
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_cancellation.id (is required)",
+		},
+		{
+			name: "Should return an error if market_id is not provided",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionCancellation{
+					Id:       "1",
+					MarketId: "",
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_cancellation.market_id (is required)",
+		},
+		{
+			name: "Should succeed if id and market id are provided",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionCancellation{
+					Id:       "123",
+					MarketId: "abcd",
+				},
+			},
+			wantErr: assert.NoError,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotErr := commands.CheckLiquidityProvisionCancellation(tt.args.cmd)
+			tt.wantErr(t, gotErr, fmt.Sprintf("CheckLiquidityProvisionCancellation(%v)", tt.args.cmd))
+			if tt.errString != "" {
+				assert.EqualError(t, gotErr, tt.errString)
+			}
+		})
+	}
+}
+
+func TestCheckLiquidityProvisionAmendment(t *testing.T) {
+	type args struct {
+		cmd *commandspb.LiquidityProvisionAmendment
+	}
+	tests := []struct {
+		name      string
+		args      args
+		wantErr   assert.ErrorAssertionFunc
+		errString string
+	}{
+		{
+			name: "Should return an error when the command is nil",
+			args: args{
+				cmd: nil,
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_amendment (is required)",
+		},
+		{
+			name: "Should return an error when id is not provided",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "",
+					MarketId: "abcd",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: 10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: -10, Proportion: 1},
+					},
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_amendment.id (is required)",
+		},
+		{
+			name: "Should return an error when market_id is not provided",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: 10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: -10, Proportion: 1},
+					},
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_amendment.market_id (is required)",
+		},
+		{
+			name: "Should return an error if amendment changes nothing",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_amendment (is required)",
+		},
+		{
+			name: "Should return no errors if amendment buys and sells are balanced",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: 10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: -10, Proportion: 1},
+					},
+				},
+			},
+			wantErr: assert.NoError,
+		},
+		{
+			name: "Should not return an error if sell side shape is provided with no buy side shape",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: 10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{},
+				},
+			},
+			wantErr: assert.NoError,
+		},
+		{
+			name: "Should not return an error if buy side shape is provided with no sell side shape",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+					Sells:    []*types.LiquidityOrder{},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: -10, Proportion: 1},
+					},
+				},
+			},
+			wantErr: assert.NoError,
+		},
+		{
+			name: "Should return errors if shapes are provided with invalid reference prices",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: 10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: -10, Proportion: 1},
+					},
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_submission.buys.0.reference (order in buy side shape with best ask price reference), liquidity_provision_submission.sells.0.offset (order in sell side shape with best bid price reference)",
+		},
+		{
+			name: "Liquidity Provision shapes with multiple errors should return all errors found",
+			args: args{
+				cmd: &commandspb.LiquidityProvisionAmendment{
+					Id:       "123",
+					MarketId: "abcd",
+					Sells: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_MID, Offset: 0, Proportion: 1},
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_BID, Offset: 0, Proportion: 1},
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_BEST_ASK, Offset: -10, Proportion: 1},
+					},
+					Buys: []*types.LiquidityOrder{
+						{Reference: types.PeggedReference_PEGGED_REFERENCE_MID, Offset: 0, Proportion: 1},
+					},
+				},
+			},
+			wantErr:   assert.Error,
+			errString: "liquidity_provision_submission.buys.0.offset (order in buy side shape offset must be < 0), liquidity_provision_submission.sells.0.offset (order in sell shape offset must be > 0), liquidity_provision_submission.sells.1.offset (order in sell side shape with best bid price reference), liquidity_provision_submission.sells.2.offset (order in sell shape offset must be >= 0)",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotErr := commands.CheckLiquidityProvisionAmendment(tt.args.cmd)
+			tt.wantErr(t, gotErr, fmt.Sprintf("CheckLiquidityProvisionAmendment(%v)", tt.args.cmd))
+
+			if tt.errString != "" {
+				assert.EqualError(t, gotErr, tt.errString)
+			}
+		})
 	}
 }