+++ vega/commands/v1/transaction.validator.pb.go
@@ -80,6 +80,20 @@ func (this *InputData) Validate() error {
 			}
 		}
 	}
+	if oneOfNester, ok := this.GetCommand().(*InputData_LiquidityProvisionCancellation); ok {
+		if oneOfNester.LiquidityProvisionCancellation != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.LiquidityProvisionCancellation); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("LiquidityProvisionCancellation", err)
+			}
+		}
+	}
+	if oneOfNester, ok := this.GetCommand().(*InputData_LiquidityProvisionAmendment); ok {
+		if oneOfNester.LiquidityProvisionAmendment != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.LiquidityProvisionAmendment); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("LiquidityProvisionAmendment", err)
+			}
+		}
+	}
 	if oneOfNester, ok := this.GetCommand().(*InputData_NodeRegistration); ok {
 		if oneOfNester.NodeRegistration != nil {
 			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.NodeRegistration); err != nil {