+++ vega/events/v1/events.validator.pb.go
@@ -111,6 +111,20 @@ func (this *TxErrorEvent) Validate() error {
 			}
 		}
 	}
+	if oneOfNester, ok := this.GetTransaction().(*TxErrorEvent_LiquidityProvisionCancellation); ok {
+		if oneOfNester.LiquidityProvisionCancellation != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.LiquidityProvisionCancellation); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("LiquidityProvisionCancellation", err)
+			}
+		}
+	}
+	if oneOfNester, ok := this.GetTransaction().(*TxErrorEvent_LiquidityProvisionAmendment); ok {
+		if oneOfNester.LiquidityProvisionAmendment != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.LiquidityProvisionAmendment); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("LiquidityProvisionAmendment", err)
+			}
+		}
+	}
 	return nil
 }
 func (this *TimeUpdate) Validate() error {
@@ -416,6 +430,13 @@ func (this *BusEvent) Validate() error {
 		if oneOfNester.StateVar != nil {
 			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.StateVar); err != nil {
 				return github_com_mwitkow_go_proto_validators.FieldError("StateVar", err)
+			}
+		}
+	}
+	if oneOfNester, ok := this.GetEvent().(*BusEvent_NetworkLimits); ok {
+		if oneOfNester.NetworkLimits != nil {
+			if err := github_com_mwitkow_go_proto_validators.CallValidatorIfExists(oneOfNester.NetworkLimits); err != nil {
+				return github_com_mwitkow_go_proto_validators.FieldError("NetworkLimits", err)
 			}
 		}
 	}